requirements.txt
python-dotenv
litellm
docker
sqlalchemy
rich
# For Image Generation
diffusers
transformers
accelerate
torch
# Provider-specific SDKs
mistralai
groq
# For the Rich Terminal UI
textual
